refactor(ImperialVehicle): fix according to code review
