+++ Content.Shared/Imperial/ImperialVehicle/Enums.cs
@@ -1,0 +1,11 @@
+using Robust.Shared.Serialization;
+
+namespace Content.Shared.Imperial.ImperialVehicle.Enums;
+
+[Serializable, NetSerializable]
+public enum VehicleVisuals : byte
+{
+    DrawDepth,
+    AutoAnimate,
+    HideRider
+}
