+++ Content.Shared/Imperial/ImperialVehicle/Events/HornActionEvent.cs
@@ -1,0 +1,5 @@
+using Content.Shared.Actions;
+
+namespace Content.Shared.Imperial.ImperialVehicle.Events;
+
+public sealed partial class HornActionEvent : InstantActionEvent;
